requirements.txt
pydantic==2.6.3
streamlit==1.32.0
pillow==10.2.0